+++ Single Mode/Single Mode/Header.h
@@ -34,7 +34,7 @@ snake* Load(int*, int*, COORD*, int*);
 void Save(int*, int*, COORD*, int*, snake*);
 void CreateMap();
 void SetColor(unsigned short);
-void Pos(int, int);
+void Pos(short,short);
 void HideCursor();
 void PrintRules(int *);
 int GameOver(int, int);
